new find algorithm

+++ Utils.h
@@ -26,7 +26,7 @@ namespace utils {
     template<typename Function, typename ...Args>
     std::tuple<double, double, double> runAndCheckTime(int numberOfRuns, const Function &fn) {
         std::vector<long long> results;
-        long long  max_result = 0;
+        long long max_result = 0;
         long long min_result = 999999999999;
         for (int i = 0; i < numberOfRuns; i++) {
             auto start = std::chrono::high_resolution_clock::now();
@@ -34,8 +34,8 @@ namespace utils {
             auto end = std::chrono::high_resolution_clock::now();
 
             results.push_back((end - start).count());
-            max_result = std::max(max_result, (end - start).count());
-            min_result = std::min(min_result, (end - start).count());
+            max_result = std::max(max_result, static_cast<long long>((end - start).count()));
+            min_result = std::min(min_result, static_cast<long long>((end - start).count()));
 //            std::cout << results.back() / 1'000'000'000. << " s " << std::endl;
         }
         return std::make_tuple(std::reduce(results.begin(), results.end()) / results.size(),
